requirements.txt
kopf
kubernetes
pykube-ng
pyyaml
jinja2
